Test: connected treeView with tableView. Working.
To make it work, I needed to remove the filterAcceptsColumn() function
from the table proxy model. It seems that mapFromSource(), which is
called in this function, returns an invalid index when using
filterAcceptsColumn() in the model. I already reported a bug[1] about
this and asked in the lists. I am waiting for an answer.

[1]: http://bugreports.qt.nokia.com/browse/QTBUG-18215