Connected main table view with table proxy model.
I connected the main table view with the new table proxy model, whichs
model is set to the dom model. For the moment, the table view is just
displaying the tasks of the current day. Other functionality is not
working but it will be migrated from the original tabel model within the
next days.